backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.5
# stability pins
setuptools>=70,<83
wheel>=0.40
pip>=23,<26